Fixed duplicate symbols error
