+++ source/Constants.cpp
@@ -1,0 +1,24 @@
+//
+//  Constants.cpp
+//  SweetSweetBetrayal
+//
+//  Created by Caitlyn Jin on 3/3/25.
+//
+#include "Constants.h"
+#include "string"
+
+namespace Constants {
+
+/**
+ * Convert an Item enum to the corresponding string.
+ */
+std::string itemToString(Item item) {
+    switch (item) {
+        case PLATFORM:
+            return "platform";
+        default:
+            return "unknown";
+    }
+}
+
+}
